Revert "Read Report Data Success. Report datas will display on web console"
This reverts commit 620ed4bc711a724d6b3b2c7f599b982d1f317640.

# Conflicts:
#	Driver/GloveDriverDLL/glove.cpp

Parse report data finished.
Now we can see the rotation of vive controller in SteamVR .

+++ Driver/GloveDriverDLL/MadgwickAHRS.h
@@ -24,14 +24,14 @@
 //----------------------------------------------------------------------------------------------------
 // Variable declaration
 
-extern volatile double beta;				// algorithm gain
-extern volatile double q0, q1, q2, q3;	// quaternion of sensor frame relative to auxiliary frame
+extern volatile float beta;				// algorithm gain
+extern volatile float q0, q1, q2, q3;	// quaternion of sensor frame relative to auxiliary frame
 
 //---------------------------------------------------------------------------------------------------
 // Function declarations
 
 void MadgwickAHRSupdate(PGLOVE_POSE_DATA_T resultData);
-void MadgwickAHRSupdateIMU(double gx, double gy, double gz, double ax, double ay, double az);
+void MadgwickAHRSupdateIMU(float gx, float gy, float gz, float ax, float ay, float az);
 
 #endif
 //=====================================================================================================
